# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CivicConnect v2 API",
    "description": "Backend API for CivicConnect v2 (design section 51A contract)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Create Incident",
        "description": "§51A.14: groups related cases under a geographic boundary. Ward officers and above.",
        "operationId": "create_incident_api_v1_incidents_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client generated UUID; a retry with the same key replays the first response (§51A.1).",
              "title": "Idempotency-Key"
            },
            "description": "Client generated UUID; a retry with the same key replays the first response (§51A.1)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_api_v1_incidents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_IncidentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analytics/overview": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Overview",
        "description": "§51A.13 aggregate operational metrics for the caller's scope (staff: their scope; overlooker: whole city, aggregates only).",
        "operationId": "analytics_overview_api_v1_analytics_overview_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 7,
              "default": 30,
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/departments/{department_id}": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Department",
        "operationId": "analytics_department_api_v1_analytics_departments__department_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "department_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Department Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/departments": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Departments",
        "operationId": "analytics_departments_api_v1_analytics_departments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/wards": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Wards",
        "operationId": "analytics_wards_api_v1_analytics_wards_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/hotspots": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Hotspots",
        "operationId": "analytics_hotspots_api_v1_analytics_hotspots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/recurrence": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Recurrence",
        "operationId": "analytics_recurrence_api_v1_analytics_recurrence_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/analytics/incidents": {
      "get": {
        "tags": [
          "analytics"
        ],
        "summary": "Analytics Incidents",
        "operationId": "analytics_incidents_api_v1_analytics_incidents_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/explain": {
      "post": {
        "tags": [
          "analytics"
        ],
        "summary": "Explain Analytics",
        "description": "AI-5 (additive, not in §51A): plain-language explanation of deterministic analytics facts. Numbers come from ``facts``; the prose is verified against them.",
        "operationId": "explain_analytics_api_v1_analytics_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsExplainResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsExplainRequest": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/AnalyticsScopeIn"
          }
        },
        "type": "object",
        "title": "AnalyticsExplainRequest"
      },
      "AnalyticsExplainResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "highlights": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsHighlightOut"
            },
            "type": "array",
            "title": "Highlights"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "type": "array",
            "title": "Citations"
          },
          "facts": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Facts"
          },
          "grounded": {
            "type": "boolean",
            "title": "Grounded"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "ai_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Metadata"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "highlights",
          "citations",
          "facts",
          "grounded"
        ],
        "title": "AnalyticsExplainResponse",
        "description": "Numbers come from ``facts`` (deterministic); ``summary``/``highlights`` only explain them and are verified against them (``grounded``)."
      },
      "AnalyticsHighlightOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "fact_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fact Ids"
          }
        },
        "type": "object",
        "required": [
          "text",
          "fact_ids"
        ],
        "title": "AnalyticsHighlightOut"
      },
      "AnalyticsScopeIn": {
        "properties": {
          "ward_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ward Id"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          }
        },
        "type": "object",
        "title": "AnalyticsScopeIn"
      },
...
      "CitationOut": {
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "CASE",
              "ANALYTIC",
              "INCIDENT"
            ],
            "title": "Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id"
        ],
        "title": "CitationOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncidentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 5000,
            "title": "Description",
            "default": ""
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "boundary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Boundary"
          },
          "case_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Case Ids"
          }
        },
        "type": "object",
        "required": [
          "title",
          "boundary"
        ],
        "title": "IncidentCreate"
      },
...
      "IncidentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "boundary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary"
          },
          "ward_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ward Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "case_count": {
            "type": "integer",
            "title": "Case Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "started_at"
        ],
        "title": "IncidentOut"
      },
...
      "Page_IncidentOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentOut"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[IncidentOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}